Move RedisType to options (#170)
